network: move typedefs around and fix includes
