+++ pkg/ebpf/c/vmlinux.h
@@ -1,6 +1,8 @@
 #ifndef __VMLINUX_H__
 #define __VMLINUX_H__
 
+;
+; // don't remove: clangd parsing bug https://github.com/clangd/clangd/issues/1167
 #pragma clang attribute push(__attribute__((preserve_access_index)), apply_to = record)
 
 typedef signed char __s8;
